# docs/openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce Mock API
  description: 이커머스 시스템 설계 및 Mock API 문서
  version: v1.0.0
servers:
- url: http://localhost:8081
  description: Generated server url
tags:
- name: Order
  description: 주문 API
- name: Coupon
  description: 쿠폰 API
- name: Product
  description: 상품 API
- name: Point
  description: 포인트 API
paths:
  /api/points/users/{userId}/charge:
    post:
      tags:
      - Point
      summary: 포인트 충전
      description: 사용자의 포인트를 충전합니다.
      operationId: chargePoint
      parameters:
      - name: userId
        in: path
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        description: 충전 요청
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargeRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseChargeResponse'
  /api/orders:
    post:
      tags:
      - Order
      summary: 주문 생성
      description: 새로운 주문을 생성합니다.
      operationId: createOrder
      requestBody:
        description: 주문 생성 요청
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseResponse'
  /api/orders/{orderId}/payment:
    post:
      tags:
      - Order
      summary: 결제 처리
      description: 주문에 대한 결제를 처리합니다.
      operationId: processPayment
      parameters:
      - name: orderId
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      requestBody:
        description: 결제 요청
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePaymentResponse'
  /api/coupons/{couponId}/issue:
    post:
      tags:
      - Coupon
      summary: 쿠폰 발급
      description: 사용자에게 쿠폰을 발급합니다.
      operationId: issueCoupon
      parameters:
      - name: couponId
        in: path
        description: 쿠폰 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: userId
        in: query
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserCouponDto'
  /api/products:
    get:
      tags:
      - Product
      summary: 상품 목록 조회
      description: 전체 상품 목록을 조회합니다.
      operationId: getProducts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListResponse'
  /api/products/{productId}:
    get:
      tags:
      - Product
      summary: 상품 상세 조회
      description: 특정 상품의 상세 정보를 조회합니다.
      operationId: getProduct
      parameters:
      - name: productId
        in: path
        description: 상품 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseResponse'
  /api/products/top:
    get:
      tags:
      - Product
      summary: 인기 상품 Top 5 조회
      description: 최근 3일간 판매량 기준 인기 상품 상위 5개를 조회합니다.
      operationId: getTopProducts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListResponse'
  /api/points/users/{userId}/history:
    get:
      tags:
      - Point
      summary: 포인트 이력 조회
      description: 사용자의 포인트 이력을 조회합니다.
      operationId: getHistory
      parameters:
      - name: userId
        in: path
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePointHistoryListResponse'
  /api/points/users/{userId}/balance:
    get:
      tags:
      - Point
      summary: 포인트 잔액 조회
      description: 사용자의 포인트 잔액을 조회합니다.
      operationId: getBalance
      parameters:
      - name: userId
        in: path
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBalanceResponse'
  /api/orders/{orderId}:
    get:
      tags:
      - Order
      summary: 주문 조회
      description: 특정 주문의 상세 정보를 조회합니다.
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseResponse'
  /api/orders/users/{userId}:
    get:
      tags:
      - Order
      summary: 사용자 주문 목록 조회
      description: 특정 사용자의 주문 목록을 조회합니다.
      operationId: getUserOrders
      parameters:
      - name: userId
        in: path
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListResponse'
  /api/coupons/users/{userId}:
    get:
      tags:
      - Coupon
      summary: 사용자 쿠폰 목록 조회
      description: 특정 사용자의 쿠폰 목록을 조회합니다.
      operationId: getUserCoupons
      parameters:
      - name: userId
        in: path
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserCouponListResponse'
components:
  schemas:
    ChargeRequest:
      required:
      - amount
      type: object
      properties:
        amount:
          type: integer
          description: 충전 금액
          format: int32
          example: 10000
      description: 포인트 충전 요청
    ApiResponseChargeResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/ChargeResponse'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    ChargeResponse:
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        chargedAmount:
          type: integer
          description: 충전 금액
          format: int32
          example: 10000
        currentBalance:
          type: integer
          description: 현재 잔액
          format: int32
          example: 60000
        chargedAt:
          type: string
          description: 충전 일시
          format: date-time
      description: 포인트 충전 응답
    CreateRequest:
      required:
      - items
      - userId
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        items:
          type: array
          description: 주문 상품 목록
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        userCouponId:
          type: integer
          description: 사용자 쿠폰 ID
          format: int64
          example: 1
      description: 주문 생성 요청
    OrderItemRequest:
      required:
      - productId
      - quantity
      type: object
      properties:
        productId:
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        quantity:
          type: integer
          description: 수량
          format: int32
          example: 2
      description: 주문 상품 요청
    ApiResponseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/Response'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    OrderItemResponse:
      type: object
      properties:
        productId:
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        productName:
          type: string
          description: 상품명
          example: 무선 키보드
        quantity:
          type: integer
          description: 수량
          format: int32
          example: 2
        price:
          type: integer
          description: 단가
          format: int32
          example: 89000
        subtotal:
          type: integer
          description: 소계
          format: int32
          example: 178000
      description: 주문 상품 응답
    Response:
      type: object
      properties:
        orderId:
          type: integer
          description: 주문 ID
          format: int64
          example: 100
        orderNumber:
          type: string
          description: 주문 번호
          example: ORD-20241029-001
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        items:
          type: array
          description: 주문 상품 목록
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        totalAmount:
          type: integer
          description: 총 금액
          format: int32
          example: 178000
        discountAmount:
          type: integer
          description: 할인 금액
          format: int32
          example: 5000
        finalAmount:
          type: integer
          description: 최종 결제 금액
          format: int32
          example: 173000
        status:
          type: string
          description: 주문 상태
          example: PENDING
        createdAt:
          type: string
          description: 주문 일시
          format: date-time
      description: 주문 응답
    PaymentRequest:
      type: object
      properties:
        usePoint:
          type: integer
          description: 사용할 포인트
          format: int32
          example: 10000
      description: 결제 요청
    ApiResponsePaymentResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/PaymentResponse'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    PaymentResponse:
      type: object
      properties:
        orderId:
          type: integer
          description: 주문 ID
          format: int64
          example: 100
        paymentStatus:
          type: string
          description: 결제 상태
          example: COMPLETED
        paymentAmount:
          type: integer
          description: 결제 금액
          format: int32
          example: 163000
        usedPoint:
          type: integer
          description: 사용 포인트
          format: int32
          example: 10000
        paymentAt:
          type: string
          description: 결제 일시
          format: date-time
      description: 결제 응답
    ApiResponseUserCouponDto:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/UserCouponDto'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    CouponDto:
      type: object
      properties:
        couponId:
          type: integer
          description: 쿠폰 ID
          format: int64
          example: 1
        name:
          type: string
          description: 쿠폰명
          example: 10% 할인 쿠폰
        discountType:
          type: string
          description: 할인 타입
          example: PERCENTAGE
        discountValue:
          type: integer
          description: 할인 값
          format: int32
          example: 10
        maxDiscountAmount:
          type: integer
          description: 최대 할인 금액
          format: int32
          example: 10000
      description: 쿠폰 정보
    UserCouponDto:
      type: object
      properties:
        userCouponId:
          type: integer
          description: 사용자 쿠폰 ID
          format: int64
          example: 1
        coupon:
          $ref: '#/components/schemas/CouponDto'
        issuedAt:
          type: string
          description: 발급 일시
          format: date-time
        expiredAt:
          type: string
          description: 만료 일시
          format: date-time
        used:
          type: boolean
          description: 사용 여부
          example: false
      description: 사용자 쿠폰 정보
    ApiResponseListResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/ListResponse'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    ListResponse:
      type: object
      properties:
        products:
          type: array
          description: 상품 목록
          items:
            $ref: '#/components/schemas/Response'
        totalCount:
          type: integer
          description: 전체 상품 수
          format: int32
          example: 5
      description: 상품 목록 응답
    ApiResponsePointHistoryListResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/PointHistoryListResponse'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    PointHistoryDto:
      type: object
      properties:
        historyId:
          type: integer
          description: 이력 ID
          format: int64
          example: 1
        transactionType:
          type: string
          description: 거래 타입
          example: CHARGE
        amount:
          type: integer
          description: 금액
          format: int32
          example: 10000
        balance:
          type: integer
          description: 잔액
          format: int32
          example: 60000
        description:
          type: string
          description: 설명
          example: 포인트 충전
        createdAt:
          type: string
          description: 거래 일시
          format: date-time
      description: 포인트 이력
    PointHistoryListResponse:
      type: object
      properties:
        histories:
          type: array
          description: 포인트 이력 목록
          items:
            $ref: '#/components/schemas/PointHistoryDto'
        totalCount:
          type: integer
          description: 전체 이력 수
          format: int32
          example: 10
      description: 포인트 이력 목록 응답
    ApiResponseBalanceResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/BalanceResponse'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    BalanceResponse:
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        balance:
          type: integer
          description: 현재 잔액
          format: int32
          example: 50000
      description: 포인트 잔액 응답
    ApiResponseUserCouponListResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/UserCouponListResponse'
        message:
          type: string
          description: 메시지
          example: 요청이 성공적으로 처리되었습니다.
      description: 공통 API 응답
    UserCouponListResponse:
      type: object
      properties:
        coupons:
          type: array
          description: 사용자 쿠폰 목록
          items:
            $ref: '#/components/schemas/UserCouponDto'
        totalCount:
          type: integer
          description: 전체 쿠폰 수
          format: int32
          example: 3
      description: 사용자 쿠폰 목록 응답